Extract `printSequence`
