+++ src/language-css/printer-postcss.js
@@ -1,7 +1,5 @@
 import printNumber from "../utils/print-number.js";
 import printString from "../utils/print-string.js";
-import hasNewline from "../utils/has-newline.js";
-import isFrontMatter from "../utils/front-matter/is-front-matter.js";
 import isNextLineEmpty from "../utils/is-next-line-empty.js";
 import isNonEmptyArray from "../utils/is-non-empty-array.js";
 import {
@@ -50,6 +48,7 @@ import {
 import { locStart, locEnd } from "./loc.js";
 import printUnit from "./utils/print-unit.js";
 import printCommaSeparatedValueGroup from "./print/comma-separated-value-group.js";
+import printSequence from "./sequence.js";
 
 function shouldPrintComma(options) {
   return options.trailingComma === "es5" || options.trailingComma === "all";
@@ -62,7 +61,7 @@ function genericPrint(path, options, print) {
     case "front-matter":
       return [node.raw, hardline];
     case "css-root": {
-      const nodes = printNodeSequence(path, options, print);
+      const nodes = printSequence(path, options, print);
       let after = node.raws.after.trim();
       if (after.startsWith(";")) {
         after = after.slice(1).trim();
@@ -96,7 +95,7 @@ function genericPrint(path, options, print) {
                 : "",
               "{",
               node.nodes.length > 0
-                ? indent([hardline, printNodeSequence(path, options, print)])
+                ? indent([hardline, printSequence(path, options, print)])
                 : "",
               hardline,
               "}",
@@ -673,49 +672,6 @@ function genericPrint(path, options, print) {
       /* c8 ignore next */
       throw new UnexpectedNodeError(node, "PostCSS");
   }
-}
-
-function printNodeSequence(path, options, print) {
-  const parts = [];
-  path.each(() => {
-    const { node, previous } = path;
-    if (
-      previous?.type === "css-comment" &&
-      previous.text.trim() === "prettier-ignore"
-    ) {
-      parts.push(options.originalText.slice(locStart(node), locEnd(node)));
-    } else {
-      parts.push(print());
-    }
-
-    if (path.isLast) {
-      return;
-    }
-
-    const { next } = path;
-    if (
-      (next.type === "css-comment" &&
-        !hasNewline(options.originalText, locStart(next), {
-          backwards: true,
-        }) &&
-        !isFrontMatter(node)) ||
-      (next.type === "css-atrule" &&
-        next.name === "else" &&
-        node.type !== "css-comment")
-    ) {
-      parts.push(" ");
-    } else {
-      parts.push(options.__isHTMLStyleAttribute ? line : hardline);
-      if (
-        isNextLineEmpty(options.originalText, locEnd(node)) &&
-        !isFrontMatter(node)
-      ) {
-        parts.push(hardline);
-      }
-    }
-  }, "nodes");
-
-  return parts;
 }
 
 const STRING_REGEX = /(["'])(?:(?!\1)[^\\]|\\.)*\1/gs;
